Refactor server startup
