+++ src/lib.rs
@@ -10,6 +10,7 @@ pub mod control;
 pub mod peers;
 pub mod retention;
 pub mod storage;
+pub mod server;
 pub mod wildmat;
 
 #[derive(Default)]
